third_party/WebKit/LayoutTests/inspector/tracing/timeline-layout-invalidations-expected.txt

Tests the Timeline API instrumentation of layout invalidations.

PASS


Running: testLocalFrame

Running: testSubframe
